Made some fixes. Moved Matrix Funcs to rtllib. Renamed Maximum, Minimum, Summation.
